requirements.txt
pandas
numpy
scikit-learn
joblib
psycopg2-binary
python-dotenv
fastapi       # only if you want API later
uvicorn       # only if you want API later
supabase
pydantic
python-jose
boto3
jsonschema>=4.0.0